Update patch set 2
Patch Set 2: Verified+1 Code-Review-2

Though it is functional, this has not been cleaned up sufficiently to actually push. Dave Daynard may clean it up over the weekend, otherwise I will on Tuesday.

What this does at the moment, is simply prevent calls for cancelAutoFocus from reaching the vendor blob, since that crashes the camera.

One additional change is required along with this, which is to rename the vendor blob to vendor-camera.msm8960.so

Patch-set: 2