Update patch set 2
Patch Set 2:

One problem is the hackish change to TheMuppets required to implement this..the vendor directory scripts come from proprietaryfiles.txt (which you need to update in this commit as well) and are not supposed to be hand-made

I propose putting the wrapper script in the device tree but not building it, and using a pre-compiled blob instead at build time. Doing so will not require any changes to d2.

Patch-set: 2