Update patch set 1
Patch Set 1:

Adam: Orientation awareness doesn't change the behavior that the display switches to landscape when the keyboard is out. That still remains.

But there are apps that enforce portrait orientation, even when the keyboard is out. With orientation aware keyboard, the cursor keys work correctly in both landscape and portrait orientations.

If you insist on orientation non-aware keyboard with keymap valid only for landscape, the cursor keys will be correct only for landscape capable apps.

For example, try Appolo with keyboard out and go to a song listing for an album (that view enforces portrait). On Photon Q (with keyboard specified as orientation aware from the beginning), the function of cursor keys follows the display orientation and it's always correct.

The real issue here is that the DPAD keys in https://github.com/CyanogenMod/android_device_samsung_apexqtmo/blob/cm-10.2/keyboard/sec_keypad.kl needs to be specified from the portrait point of view, so that the cursor keys work correctly as orientation aware.

Patch-set: 1