Update patch set 1
Patch Set 1:

Tom, for any changes that you want merged to stable, please open up the change request (you can use the cherry-pick option, or push it as a new change), or add it to the spreadsheet. Otherwise, there is a decent chance it will get lost in the mix.

I double check everything that is pending for the stable/ branches in gerrit and everything that is listed on the notes for the spreadsheet before I kick the builds.

Patch-set: 1